AIAgent_Scratch/pyproject.toml
[project]
name = "aiagent-scratch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "groq>=1.0.0",
    "python-dotenv>=1.2.1",
]
